requirements.txt
Django==4.2.7
python-decouple==3.8
django-crispy-forms==2.1
crispy-bootstrap5==0.7
psycopg2-binary==2.9.9
dj-database-url==2.1.0
whitenoise==6.6.0
python-dotenv==1.0.1

boto3==1.34.29
django-storages==1.14.2
Pillow==10.1.0
